+++ src/Validator/Rules/OverlappingFieldsCanBeMerged.php
@@ -44,7 +44,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @var PairSet
      */
-    private $comparedFragmentPairs;
+    protected $comparedFragmentPairs;
 
     /**
      * A cache for the "field map" and list of fragment names found in any given
@@ -53,7 +53,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @var SplObjectStorage
      */
-    private $cachedFieldsAndFragmentNames;
+    protected $cachedFieldsAndFragmentNames;
 
     public function getVisitor(ValidationContext $context)
     {
@@ -89,7 +89,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @return mixed[]
      */
-    private function findConflictsWithinSelectionSet(
+    protected function findConflictsWithinSelectionSet(
         ValidationContext $context,
         $parentType,
         SelectionSetNode $selectionSet
@@ -152,7 +152,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @return mixed[]|SplObjectStorage
      */
-    private function getFieldsAndFragmentNames(
+    protected function getFieldsAndFragmentNames(
         ValidationContext $context,
         $parentType,
         SelectionSetNode $selectionSet
@@ -239,7 +239,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * @param mixed[][][]   $astAndDefs
      * @param bool[]        $fragmentNames
      */
-    private function internalCollectFieldsAndFragmentNames(
+    protected function internalCollectFieldsAndFragmentNames(
         ValidationContext $context,
         $parentType,
         SelectionSetNode $selectionSet,
@@ -296,7 +296,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * @param mixed[][] $conflicts
      * @param mixed[][] $fieldMap
      */
-    private function collectConflictsWithin(
+    protected function collectConflictsWithin(
         ValidationContext $context,
         array &$conflicts,
         array $fieldMap
@@ -344,7 +344,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @return mixed[]|null
      */
-    private function findConflict(
+    protected function findConflict(
         ValidationContext $context,
         $parentFieldsAreMutuallyExclusive,
         $responseName,
@@ -433,7 +433,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * @param NodeList<ArgumentNode> $arguments1 keep
      * @param NodeList<ArgumentNode> $arguments2 keep
      */
-    private function sameArguments(NodeList $arguments1, NodeList $arguments2): bool
+    protected function sameArguments(NodeList $arguments1, NodeList $arguments2): bool
     {
         if (count($arguments1) !== count($arguments2)) {
             return false;
@@ -463,7 +463,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
     /**
      * @return bool
      */
-    private function sameValue(Node $value1, Node $value2)
+    protected function sameValue(Node $value1, Node $value2)
     {
         return (! $value1 && ! $value2) || (Printer::doPrint($value1) === Printer::doPrint($value2));
     }
@@ -473,7 +473,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * Composite types are ignored as their individual field types will be compared
      * later recursively. However List and Non-Null types must match.
      */
-    private function doTypesConflict(Type $type1, Type $type2): bool
+    protected function doTypesConflict(Type $type1, Type $type2): bool
     {
         if ($type1 instanceof ListOfType) {
             return $type2 instanceof ListOfType
@@ -517,7 +517,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @return mixed[][]
      */
-    private function findConflictsBetweenSubSelectionSets(
+    protected function findConflictsBetweenSubSelectionSets(
         ValidationContext $context,
         $areMutuallyExclusive,
         $parentType1,
@@ -611,7 +611,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * @param mixed[]   $fieldMap1
      * @param mixed[]   $fieldMap2
      */
-    private function collectConflictsBetween(
+    protected function collectConflictsBetween(
         ValidationContext $context,
         array &$conflicts,
         $parentFieldsAreMutuallyExclusive,
@@ -660,7 +660,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * @param mixed[][] $fieldMap
      * @param string    $fragmentName
      */
-    private function collectConflictsBetweenFieldsAndFragment(
+    protected function collectConflictsBetweenFieldsAndFragment(
         ValidationContext $context,
         array &$conflicts,
         array &$comparedFragments,
@@ -719,7 +719,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @return mixed[]|SplObjectStorage
      */
-    private function getReferencedFieldsAndFragmentNames(
+    protected function getReferencedFieldsAndFragmentNames(
         ValidationContext $context,
         FragmentDefinitionNode $fragment
     ) {
@@ -746,7 +746,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      * @param string    $fragmentName1
      * @param string    $fragmentName2
      */
-    private function collectConflictsBetweenFragments(
+    protected function collectConflictsBetweenFragments(
         ValidationContext $context,
         array &$conflicts,
         $areMutuallyExclusive,
@@ -836,7 +836,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule
      *
      * @return mixed[]|null
      */
-    private function subfieldConflicts(
+    protected function subfieldConflicts(
         array $conflicts,
         $responseName,
         FieldNode $ast1,