+++ src/Validator/Rules/ValuesOfCorrectType.php
@@ -187,7 +187,7 @@ class ValuesOfCorrectType extends ValidationRule
     /**
      * @param VariableNode|NullValueNode|IntValueNode|FloatValueNode|StringValueNode|BooleanValueNode|EnumValueNode|ListValueNode|ObjectValueNode $node
      */
-    private function isValidScalar(ValidationContext $context, ValueNode $node, $fieldName)
+    protected function isValidScalar(ValidationContext $context, ValueNode $node, $fieldName)
     {
         // Report any error at the full type expected by the location.
         /** @var ScalarType|EnumType|InputObjectType|ListOfType|NonNull $locationType */
@@ -244,7 +244,7 @@ class ValuesOfCorrectType extends ValidationRule
     /**
      * @param VariableNode|NullValueNode|IntValueNode|FloatValueNode|StringValueNode|BooleanValueNode|EnumValueNode|ListValueNode|ObjectValueNode $node
      */
-    private function enumTypeSuggestion($type, ValueNode $node)
+    protected function enumTypeSuggestion($type, ValueNode $node)
     {
         if ($type instanceof EnumType) {
             $suggestions = Utils::suggestionList(
@@ -278,7 +278,7 @@ class ValuesOfCorrectType extends ValidationRule
             ($message ? sprintf('; %s', $message) : '.');
     }
 
-    private static function getBadValueMessage($typeName, $valueName, $message = null, $context = null, $fieldName = null)
+    protected static function getBadValueMessage($typeName, $valueName, $message = null, $context = null, $fieldName = null)
     {
         if ($context) {
             $arg = $context->getArgument();